Switch to official Rust AWS Lambda runtime
This commit is a major change to the library, switching out the Go-based
third-party runtime for the official Rust AWS Lambda runtime.

Additionally, the `PhysicalResourceIdSuffixProvider` trait has been
renamed to `PhysicalResourceIdProvider` and now allows users of the
library to overwrite the entire physical resource ID, if desired. (The
default ID generation has been kept unchanged, i.e. upgrading does not
break existing custom resources because of ID mismatches.)

All dependencies have also been upgraded to their latest versions.

Please note that right now this depends on the git-repository for the
Lambda runtime, because async-support has not yet landed in an official
release. This means that cfn-resource-provider will not be released
until the `lambda` dependency has been released.